RYSK-50: Hoist card to be re-used.
