+++ packages/front-end/src/components/shared/SimpleCard/index.tsx
