+++ packages/front-end/src/components/shared/SimpleCard/types.ts
